Add ament_lint_auto dependency to package manifest.
This package is required when BUILD_TESTING is enabled but is not in the
package manifest and is causing failures on the ROS buildfarm.
I added ament_lint_common since that seems to be the practice when using
ament_lint_auto but I do not know whether that's necessary or desired in
this case.