remove obj_id for 3d localization
currently obj_id is already zero, because the 2d tracking is not accurate and 3d localication did not use this value. but zero display will confuse user in visualization. so just remove it. Once 2d tracking is good quality and 3d localization will based on 2d tracking output, and then re-add this obj_id again.